requirements.txt
fastapi>=0.95.0
uvicorn[standard]
SQLAlchemy>=1.4
asyncpg
alembic
pydantic
python-dotenv